Moved client routes into their own folder, so the client inventory logic can be separated from the web logic. phpserialize.py was failing to serialize the correct length of strings because the length calculation was taken after the conversion from unicode, but the original unicode string was given in the request body, which always forced a mismatch. Correct usage of display_x() functions imported from munkilib
