+++ app/Http/Controllers/Client/ReportController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Client;
 
 
 use App\Notifications\BrokenClient;
@@ -8,6 +8,7 @@ use App\Notifications\CheckIn;
 use App\Notifications\GlobalNotifiable;
 use App\ReportData;
 use App\User;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Notification;
@@ -17,8 +18,6 @@ use munkireport\lib\Modules;
 use munkireport\lib\Unserializer;
 use \Messages_model, \Exception;
 use munkireport\models\Hash as MunkiReportHash;
-use function xKerman\Restricted\unserialize;
-use xKerman\Restricted\UnserializeFailedException;
 
 class ReportController extends Controller
 {
@@ -101,10 +100,9 @@ class ReportController extends Controller
             $this->_register($request->post('serial'));
 
             //$req_items = unserialize($_POST['items']); //Todo: check if array
-            //$unserializer = new Unserializer($request->post('items'));
-            //$req_items = $unserializer->unserialize();
+            $unserializer = new Unserializer($request->post('items'));
+            $req_items = $unserializer->unserialize();
 
-            $req_items = unserialize($request->post('items'));
             // Reset messages for this client
             if (isset($req_items['msg'])) {
                 $msg_obj = new Messages_model();
@@ -180,8 +178,9 @@ class ReportController extends Controller
         // Way too noisy - used for debugging notifications
         // $global->notify(new CheckIn($_POST['serial']));
 
+        Storage::disk('local')->put('failure.txt', $request->get('items'));
         $use_xkerman_unserializer = true;
-
+        // xdebug_break();
         try{
             if ($use_xkerman_unserializer) {
                 $arr = unserialize($request->get('items'));
@@ -193,6 +192,7 @@ class ReportController extends Controller
         }
         catch (Exception $e){
             Log::error($e);
+
             return response(serialize(array('error' => 'Could not unserialize items')));
         }
