+++ public/assets/client_installer/payload/usr/local/munkireport/munkilib/reportcommon.py
@@ -326,19 +326,19 @@ def process(serial, items):
 
     # Decode response
     try:
-        result = unserialize(server_data.decode('utf8'))
+        result = unserialize(server_data.decode('utf-8'))
     except Exception as e:
         display_error("Could not unserialize server data: %s" % str(e))
-        display_error("Request: ", str(values))
-        display_error("Response: ", str(server_data))
+        display_error("Request: %s", str(values))
+        display_error("Response: %s", str(server_data))
         return -1
 
     if result.get("error") != "":
-        display_error("Server error: ", result["error"])
+        display_error("Server error: %s", result["error"])
         return -1
 
     if result.get("info") != "":
-        display_detail("Server info: ", result['info'])
+        display_detail("Server info: %s", result['info'])
 
     # Retrieve hashes that need updating
     total_size = 0
